exercise Custom Hooks 03 done
